Add an interface for adding dimensions to metrics
Also add a single implementation, InstanceIdAdder, which can either add
a fixed instance id or fetch the instance id if running in EC2. The
InstanceIdAdder takes a MetricPredicate, which allows it to only add the
InstanceId dimension to instance-specific metrics.

Finally, add a builder for CloudWatchReporter to keep from exposing
a gigantic constructor to users.