Trim values to range supported by CloudWatch
Amazon only allows values smaller than 1E108 and larger than 1E-108.
This trims values outside that range to the edge and logs a debug
message when it happens for the first time. Because infrequent events
can cause rates to creep down into this range, I didn't think a warning
log was appropriate.